remove duplications and minor changes
