--- relay_rpc/src/rpc/params/params.rs
@@ -1,20 +1,0 @@
-pub mod session;
-
-use serde::{Deserialize, Serialize};
-
-#[derive(Debug, Serialize, PartialEq, Eq, Deserialize, Clone, Default)]
-#[serde(rename_all = "camelCase")]
-pub struct Metadata {
-    pub description: String,
-    pub url: String,
-    pub icons: Vec<String>,
-    pub name: String,
-}
-
-#[derive(Debug, Serialize, PartialEq, Eq, Deserialize, Clone, Default)]
-pub struct Relay {
-    pub protocol: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    #[serde(default)]
-    pub data: Option<String>,
-}
